python/pyproject.toml
[tool.poetry]
name = "algo2024"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
python_files = "*_tests.py"

[tool.poetry.group.dev.dependencies]
pytest-cov = "^5.0.0"

